V4/fix(dropdown): fix bug with focus behaviour (#2122)
