+++ .changeset/fix-Dropdown-focus-behaviour.md
@@ -1,0 +1,5 @@
+---
+'react-magma-dom': patch
+---
+
+fix(Dropdown): Fix bug with focus behaviour
